add chainmeta
