+++ pact.cabal
@@ -45,6 +45,7 @@ library
                      , Pact.Server.API
                      , Pact.Server.Client
                      , Pact.Types.API
+                     , Pact.Types.ChainMeta
                      , Pact.Types.Command
                      , Pact.Types.Scheme
                      , Pact.Types.Exp
