+++ src/Pact/Types/Runtime.hs
@@ -40,16 +40,12 @@ module Pact.Types.Runtime
    NamespacePolicy(..), nsPolicy,
    permissiveNamespacePolicy,
    SPVSupport(..),noSPVSupport,
-   Address(..),aFrom,aTo,
-   PrivateMeta(..),pmAddress,
-   PublicMeta(..),pmChainId,pmSender,pmGasLimit,pmGasPrice,pmFee,
-   HasPlafMeta(..),
    PactContinuation(..),
-   PublicData(..), pdChainId, pdPublicMeta, pdBlockHeight, pdBlockTime,
    module Pact.Types.Lang,
    module Pact.Types.Util,
    module Pact.Types.Persistence,
-   module Pact.Types.Gas
+   module Pact.Types.Gas,
+   module Pact.Types.ChainMeta
    ) where
 
 import GHC.Generics
@@ -75,6 +71,7 @@ import qualified Data.Set as S
 import Data.Word (Word32, Word64)
 
 import Pact.Parse
+import Pact.Types.ChainMeta
 import Pact.Types.Gas
 import Pact.Types.Lang
 import Pact.Types.Orphans ()
@@ -175,10 +172,6 @@ data RefStore = RefStore {
 makeLenses ''RefStore
 instance Default RefStore where def = RefStore HM.empty HM.empty
 
-newtype EntityName = EntityName Text
-  deriving (IsString,AsString,Eq,Ord,Hashable,Serialize,NFData,ToJSON,FromJSON,Default)
-instance Show EntityName where show (EntityName t) = show t
-
 newtype PactContinuation = PactContinuation (App (Term Ref))
   deriving (Eq,Show)
 
@@ -226,72 +219,6 @@ newtype SPVSupport = SPVSupport {
 
 noSPVSupport :: SPVSupport
 noSPVSupport = SPVSupport $ \_ _ -> return $ Left $ "SPV verify not supported"
-
--- | Confidential/Encrypted addressing info, for use in metadata on privacy-supporting platforms.
-data Address = Address {
-    _aFrom :: EntityName
-  , _aTo :: S.Set EntityName
-  } deriving (Eq,Show,Ord,Generic)
-instance NFData Address
-instance Serialize Address
-instance ToJSON Address where toJSON = lensyToJSON 2
-instance FromJSON Address where parseJSON = lensyParseJSON 2
-makeLenses ''Address
-
-data PrivateMeta = PrivateMeta
-  { _pmAddress :: Maybe Address
-  } deriving (Eq,Show,Generic)
-$(makeLenses ''PrivateMeta)
-instance Default PrivateMeta where def = PrivateMeta def
-instance ToJSON PrivateMeta where toJSON = lensyToJSON 3
-instance FromJSON PrivateMeta where parseJSON = lensyParseJSON 3
-instance NFData PrivateMeta
-instance Serialize PrivateMeta
-
--- | Contains all necessary metadata for a Chainweb-style public chain.
-data PublicMeta = PublicMeta
-  { _pmChainId :: Text
-  , _pmSender :: Text
-  , _pmGasLimit :: ParsedInteger
-  , _pmGasPrice :: ParsedDecimal
-  , _pmFee :: ParsedDecimal
-  } deriving (Eq,Show,Generic)
-makeLenses ''PublicMeta
-
-instance Default PublicMeta where def = PublicMeta "" "" 0 0 0
-instance ToJSON PublicMeta where toJSON = lensyToJSON 3
-instance FromJSON PublicMeta where parseJSON = lensyParseJSON 3
-instance NFData PublicMeta
-instance Serialize PublicMeta
-
-class HasPlafMeta a where
-  getPrivateMeta :: a -> PrivateMeta
-  getPublicMeta :: a -> PublicMeta
-
-instance HasPlafMeta PrivateMeta where
-  getPrivateMeta = id
-  getPublicMeta = const def
-
-instance HasPlafMeta PublicMeta where
-  getPrivateMeta = const def
-  getPublicMeta = id
-
-instance HasPlafMeta () where
-  getPrivateMeta = const def
-  getPublicMeta = const def
-
-data PublicData = PublicData
-  { _pdPublicMeta :: PublicMeta
-  , _pdChainId :: Word32
-  , _pdBlockHeight :: Word64
-  , _pdBlockTime :: Int64
-  }
-  deriving (Show, Eq, Generic)
-makeLenses ''PublicData
-
-instance ToJSON PublicData where toJSON = lensyToJSON 3
-instance FromJSON PublicData where parseJSON = lensyParseJSON 3
-instance Default PublicData where def = PublicData def def def def
 
 -- | Interpreter reader environment, parameterized over back-end MVar state type.
 data EvalEnv e = EvalEnv {
